foodgram/requirements.txt
Django==3.2.16
Pillow==9.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
django-extra-fields==3.0.2
djoser==2.1.0
django-filter==21.1
gunicorn==20.1.0
requests==2.28.1
requests-oauthlib==1.3.1
urllib3==1.26.11
psycopg2==2.9.9
django-cors-headers==3.13.0
psycopg2-binary==2.9.3 
flake8==5.0.4